requirements.txt
Werkzeug
implicit
Flask
numpy
gevent
pillow
pandas
numpy
scipy
sklearn
pymongo[srv]
pika
scikit-surprise
torch